requirements.txt
playwright==1.50.0
google-generativeai==0.8.4
pillow==12.1.0
python-dotenv==1.0.1